FNLM: Remove NEW_TASK flag from getIntentNotificationListenerSettings
The current code has no idea if the caller getting the Intent is a UI Context or not.
The code could add a Context parameter and test if it is a UI context, but that really is not this method's responsibility.
It should be the responsibility of the caller to know if it needs to add a NEW_TASK flag to the Intent or not.